+++ pt2vhf_aprs/__init__.py
@@ -1,1 +1,2 @@
-__version__ = "1.6.19"
+__version__ = "1.6.20"
+APP_TOCALL = "APZVHF"